Initial tox test implementation
