+++ server/server.py
@@ -1,7 +1,5 @@
 from typing import Optional
 import re
-import json
-import os
 
 from pygls.server import LanguageServer
 from lsprotocol.types import (
@@ -16,14 +14,8 @@ from lsprotocol.types import (
     SemanticTokens,
 )
 
-# Read regex strings from syntax
-REGEX_KEY = {}
-with open(os.path.join('syntaxes', 'nastran.json')) as f:
-    d = json.load(f)
-for key in d['repository'].keys():
-    REGEX_KEY[key.upper()] = d['repository'][key]['match']
-
-from utils.read_docs import get_docs, SECTION_KEY
+from utils.read_docs import get_docs
+from utils.parse_file import get_section, exec_regex, REGEX_KEY
 
 class NastranLanguageServer(LanguageServer):
     """Subclass of pygls LanguageServer"""
@@ -35,49 +27,6 @@ class NastranLanguageServer(LanguageServer):
 
 # Initialize server class
 server = NastranLanguageServer("NastranLanguageServer", "v0.1")
-
-def get_section(line, lines) -> str:
-    # Determine index of current line
-    ind = lines.index(line)
-    # Strip to left of string
-    raw = line.lstrip()
-    if raw.startswith('*'):
-        return 'BULK'
-    if raw.startswith('$') and not raw.startswith('$S700'):
-        return ''
-    
-    # If special character detected, strip everything after that character
-    card = raw
-    for char in [",", "*", " ", "=", "\n", "("]:
-        if char in card:
-            card = card.split(char)[0].strip()
-    
-    # Determine relative position in file and return section if detected
-    CEND = [i for i, line in enumerate(lines) if 'CEND' in line]
-    CEND = 0 if not CEND else CEND[0]
-    BBULK = [i for i, line in enumerate(lines) if 'BEGIN BULK' in line]
-    BBULK = 0 if not BBULK else BBULK[0]
-    if CEND > 0 and BBULK > 0:
-        if ind <= CEND and any([name for name in SECTION_KEY['FMS'] if name.startswith(card)]):
-            return 'FMS'
-        elif ind <= CEND and any([name for name in SECTION_KEY['EXEC'] if name.startswith(card)]):
-            return 'EXEC'
-        elif ind > CEND and ind <= BBULK and any([name for name in SECTION_KEY['CASE'] if name.startswith(card)]):
-            return 'CASE'
-        elif ind > BBULK and card in SECTION_KEY['BULK']:
-            return 'BULK'
-
-    # Test for raw line
-    for section, cards in SECTION_KEY.items():
-        if raw.strip() in cards:
-            return section
-   
-    # Catch any partial cards
-    for section, cards in SECTION_KEY.items():
-        if any([name for name in cards if name.startswith(card)]):
-            return section
-
-    return 'BULK'
 
 @server.feature(TEXT_DOCUMENT_HOVER)
 async def hovers(params: HoverParams) -> Optional[Hover]:
@@ -101,29 +50,11 @@ async def hovers(params: HoverParams) -> Optional[Hover]:
         # Process PARAM
         logic = (params.position.character >= line.lower().index('param')) and (params.position.character <= line.lower().index('param')+len('param'))
         if logic:
-            card = re.match(REGEX_KEY[section], line)
-            if card:
-                card = card.groups()[0]
-                if not card:
-                    card = ''
-            else:
-                card = ''
+            card = exec_regex(re.match, REGEX_KEY[section], line)
         else:
-            card = re.search(REGEX_KEY['PARAM'], line)
-            if card:
-                card = card.groups()[0]
-                if not card:
-                    card = ''
-            else:
-                card = ''
+            card = exec_regex(re.search, REGEX_KEY['PARAM'], line)
     else:
-        card = re.match(REGEX_KEY[section], line)
-        if card:
-            card = card.groups()[0]
-            if not card:
-                card = ''
-        else:
-            card = ''
+        card = exec_regex(re.match, REGEX_KEY[section], line)
     # Calculate hover text
     hover_txt = get_docs(card, section=section)
     # If card is not blank and not a comment, find hover text
