+++ tox.ini
@@ -1,0 +1,8 @@
+[tox]
+env_list = py{37,38,39,310,311}
+
+[testenv]
+description = test regex patterns
+deps = pytest
+skip_install = true
+commands = python -m pytest tests
